抽取工具类
